@@ -8,82 +8,65 @@
     "test:coverage": "jest --coverage"
   },
   "dependencies": {
-    "@rails/webpacker": "^3.5.3",
-    "babel-core": "^6.25.0",
-    "babel-loader": "7.x",
-    "babel-plugin-react-transform": "^2.0.0",
-    "babel-plugin-syntax-dynamic-import": "^6.18.0",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-runtime": "^6.6.0",
-    "babel-polyfill": "^6.23.0",
-    "babel-preset-env": "^1.6.0",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-react-hmre": "^1.1.1",
-    "babel-preset-stage-0": "^6.16.0",
-    "babel-preset-stage-1": "^6.16.0",
-    "babel-preset-stage-2": "^6.5.0",
-    "babel-runtime": "^6.6.1",
+    "@babel/core": "*",
+    "@babel/plugin-proposal-class-properties": "^7.5.5",
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
+    "@babel/plugin-transform-runtime": "^7.6.0",
+    "@babel/preset-env": "^7.6.0",
+    "@babel/preset-react": "^7.0.0",
+    "@rails/webpacker": "^4.0.7",
+    "babel-loader": "8.0.6",
+    "babel-plugin-react-transform": "^3.0.0",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
     "classnames": "^2.2.5",
-    "coffee-loader": "^0.7.3",
-    "coffee-script": "^1.12.7",
-    "compression-webpack-plugin": "^1.0.0",
-    "es5-shim": "^4.5.2",
-    "es6-promise": "^3.1.2",
-    "eslint-plugin-vue": "^4.3.0",
+    "compression-webpack-plugin": "^3.0.0",
+    "es6-promise": "^4.2.8",
     "expose-loader": "^0.7.1",
     "extract-text-webpack-plugin": "^3.0.0",
-    "file-loader": "^0.11.2",
+    "file-loader": "^4.2.0",
     "glob": "^7.1.2",
-    "imports-loader": "^0.6.5",
+    "imports-loader": "^0.8.0",
     "jquery": "^3.1.1",
     "jquery-ujs": "^1.2.2",
     "js-yaml": "^3.9.0",
-    "loader-utils": "^0.2.12",
+    "loader-utils": "^1.2.3",
     "lodash": "^4.17.4",
     "object-assign-shim": "^1.0.0",
-    "path-complete-extname": "^0.1.0",
-    "prop-types": "^15.6.1",
+    "path-complete-extname": "^1.0.0",
+    "prop-types": "^15.7.2",
     "rails-erb-loader": "^5.1.0",
-    "react": "^16.4.1",
+    "react": "^16.9.0",
     "react-addons-update": "^15.6.0",
-    "react-dom": "^16.4.1",
-    "resolve-url-loader": "^2.1.0",
-    "url-loader": "^0.5.7",
-    "webpack": "^3.4.1",
-    "webpack-manifest-plugin": "^1.2.1",
+    "react-dom": "^16.9.0",
+    "resolve-url-loader": "^3.1.0",
+    "url-loader": "^2.1.0",
+    "webpack": "^4.40.2",
+    "webpack-manifest-plugin": "^2.0.4",
     "webpack-merge": "^4.1.0",
-    "webpacker-react": "^0.3.0"
+    "webpacker-react": "^1.0.0-beta.1"
   },
   "devDependencies": {
-    "babel-eslint": "^7.1.0",
-    "babel-jest": "^23.0.1",
+    "babel-eslint": "^10.0.3",
+    "babel-jest": "^24.9.0",
     "enzyme": "^3.3.0",
     "enzyme-adapter-react-16": "^1.1.1",
     "enzyme-to-json": "^3.3.4",
-    "eslint": "^3.10.1",
-    "eslint-config-airbnb": "^13.0.0",
+    "eslint": "^6.3.0",
+    "eslint-config-airbnb": "^18.0.1",
     "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^2.2.3",
-    "eslint-plugin-react": "^6.7.1",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-react": "^7.14.3",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^23.1.0",
-    "jest-enzyme": "^6.0.1",
-    "react-test-renderer": "^15.6.1",
+    "jest": "^24.9.0",
+    "jest-enzyme": "^7.1.1",
+    "react-test-renderer": "^16.9.0",
     "react-transform-hmr": "^1.0.2",
-    "sinon": "^1.17.6",
+    "sinon": "^7.4.2",
     "webpack-cli": "^3.3.0",
-    "webpack-dev-server": "^3.2.1"
+    "webpack-dev-server": "^3.8.0"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/houndci/hound.git"
-  },
-  "babel": {
-    "presets": [
-      "es2015",
-      "stage-0",
-      "react"
-    ]
   }
 }
