@@ -1,4 +1,4 @@
-import NotifyTierChange from 'components/NotifyTierChange';
 import WebpackerReact from 'webpacker-react';
+import NotifyTierChange from '../components/NotifyTierChange';
 
 WebpackerReact.setup({NotifyTierChange});
