@@ -1,4 +1,4 @@
-import UpdateAccountCreditCard from 'components/UpdateAccountCreditCard';
 import WebpackerReact from 'webpacker-react';
+import UpdateAccountCreditCard from '../components/UpdateAccountCreditCard';
 
 WebpackerReact.setup({UpdateAccountCreditCard});
