@@ -1,4 +1,4 @@
-import UpdateAccountEmail from 'components/UpdateAccountEmail';
 import WebpackerReact from 'webpacker-react';
+import UpdateAccountEmail from '../components/UpdateAccountEmail';
 
 WebpackerReact.setup({UpdateAccountEmail});
